#50 #51 #52 Feat: Waiting CRUD backend 구현
기본적인 컨트롤러 - 레포지토리 까지의 코드 구현

+++ src/main/java/kr/oshino/eataku/waiting/model/dto/requestDto/CreateWaitingRequestDto.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 public class CreateWaitingRequestDto {
 
-    private int memberNo;
-    private int restaurantNo;
+    private Long memberNo;
+    private Integer restaurantNo;
     private int partySize;
 }
